Regenerate the APIv6 SDK and fix examples

+++ apiv6/localhost_apiserver/synthetics/go/model_v202101beta1_health_moment.go
@@ -20,7 +20,7 @@ type V202101beta1HealthMoment struct {
 
 	DstIp string `json:"dstIp,omitempty"`
 
-	PacketLoss int64 `json:"packetLoss,omitempty"`
+	PacketLoss float64 `json:"packetLoss,omitempty"`
 
 	AvgLatency int64 `json:"avgLatency,omitempty"`
